vista eliminar usuario y Area usuario

+++ ingresos_egresos/src/main/java/servicios/exequiales/ingresosyegresos/ingresos_egresos/Service/EmpleadoService.java
@@ -39,8 +39,8 @@ public class EmpleadoService implements IEmpleadoService {
     }
 
     @Override
-    public void deletEmpleado(int id) {
-        empleadoRepository.deleteById((long)id);
+    public void deletEmpleado(long id) {
+        empleadoRepository.deleteById(id);
     }
