+++ ingresos_egresos/src/main/java/servicios/exequiales/ingresosyegresos/ingresos_egresos/Service/IEmpleadoService.java
@@ -10,6 +10,6 @@ public interface IEmpleadoService {
     public List<Empleado> findAll();
     public Empleado createEmpleado(Empleado empleado);
     public Empleado updateEmpleado(int id, Empleado empleado);
-    public void deletEmpleado(int id);
+    public void deletEmpleado(long id);
 
 }
